Update README, make ManifestWasm public, move back Builder constructors to toplevel
