+++ src/test/java/org/extism/chicory/sdk/PluginTest.java
@@ -8,10 +8,11 @@ public class PluginTest extends TestCase {
 
     public void testGreet() {
         var manifest =
-                Manifest.Builder
-                        .fromUrl("https://github.com/extism/plugins/releases/download/v1.1.0/greet.wasm")
-                        .build();
-        var plugin = Plugin.Builder.ofManifest(manifest).build();
+                Manifest.ofWasms(
+                        ManifestWasm.fromUrl(
+                                "https://github.com/extism/plugins/releases/download/v1.1.0/greet.wasm")
+                        .build()).build();
+        var plugin = Plugin.ofManifest(manifest).build();
         var input = "Benjamin";
         var result = new String(plugin.call("greet", input.getBytes(StandardCharsets.UTF_8)), StandardCharsets.UTF_8);
         assertEquals("Hello, Benjamin!", result);
@@ -19,11 +20,14 @@ public class PluginTest extends TestCase {
 
     public void testGreetAoT() {
         var manifest =
-                Manifest.Builder
-                        .fromUrl("https://github.com/extism/plugins/releases/download/v1.1.0/greet.wasm")
+                Manifest.ofWasms(
+                        ManifestWasm.fromUrl(
+                                "https://github.com/extism/plugins/releases/download/v1.1.0/greet.wasm")
+                                .withName("greet")
+                                .build())
                         .withOptions(new Manifest.Options().withAoT())
                         .build();
-        var plugin = Plugin.Builder.ofManifest(manifest).build();
+        var plugin = Plugin.ofManifest(manifest).build();
         var input = "Benjamin";
         var result = new String(plugin.call("greet", input.getBytes(StandardCharsets.UTF_8)), StandardCharsets.UTF_8);
         assertEquals("Hello, Benjamin!", result);
